wattwise/vertex_pipeline/requirements.txt
torch==2.6.0
torchaudio==2.6.0
torchinfo==1.8.0
torchvision==0.21.0

scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2

matplotlib==3.10.0

numpy==2.2.3

pandas==2.2.3

cdsapi==0.7.5

wandb

kfp==2.7.0
google-cloud-aiplatform==1.42.1
google-cloud-storage==2.14.0
google-auth==2.27.0
google-auth-oauthlib==1.2.0
gcsfs==2024.2.0

google-cloud-bigtable
